% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoringScripts.R
\name{makeEnrichment}
\alias{makeEnrichment}
\title{Detect enriched categories in an annotation source}
\usage{
makeEnrichment(gsyms, adj = NULL, nom = NULL)
}
\arguments{
\item{gsyms}{a set of training genes}

\item{adj}{an annotation matrix}

\item{nom}{an optional vector of term names corresponding to the column names of the annotation matrix}
}
\description{
Detect enriched categories in an annotation source
}
