% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoringScripts.R
\name{makeScore}
\alias{makeScore}
\title{Score a set of training genes based on an annotation source.}
\usage{
makeScore(AnnotationMatrix = NULL, TrainingGenes = NULL, MinPerAnno = 5,
  OR = 1.5, PVal = 0.01, MinAnno = 5, MaxAnno = 200)
}
\arguments{
\item{AnnotationMatrix}{a matrix with genes as rows and annotation variables as columns (with appropriate row/column names)}

\item{TrainingGenes}{a character vector of genes}

\item{MinPerAnno}{The minimum require number of genes annotated for an annotation to be considered}

\item{OR}{minimum Odds Ratio to be reported}

\item{PVal}{minimum p-value to be reported}

\item{MinAnno}{minimum number of annotations to select}

\item{MaxAnno}{maximum number of annotations to select}
}
\value{
a list with two elements: a named vector of Scores and a matrix of EnrichedAnnotations
}
\description{
Score a set of training genes based on an annotation source.
}
